pyproject.toml
[project]
name = "project-4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.135.2",
    "google-auth>=2.49.1",
    "greenlet>=3.3.2",
    "httpx>=0.28.1",
    "langchain>=1.2.13",
    "langchain-core>=1.2.23",
    "langchain-google-genai>=4.2.1",
    "langchain-google-vertexai>=3.2.2",
    "loguru>=0.7.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-multipart>=0.0.22",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.48",
    "uvicorn[standard]>=0.42.0",
]
